入力値制限を追加

+++ SmartAttendance/src/main/java/com/example/demo/entity/RegistKintaiEntity.java
@@ -1,6 +1,5 @@
 package com.example.demo.entity;
 
-import java.math.BigDecimal;
 import java.sql.Date;
 
 import lombok.Data;
@@ -13,7 +12,7 @@ public class RegistKintaiEntity {
 		private int start_M;
 		private int end_H;
 		private int end_M;
-		private BigDecimal total_workingTime;/*実労働時間[時間]*/
+		private double total_workingTime;/*実労働時間[時間]*/
 		private int breakTime; /*休憩時間[時間]*/
 		private int excessTime; /*超過労働時間[時間]*/
